style 코드 정리
