+++ src/containers/ChoiceContainer.jsx
@@ -8,41 +8,13 @@ import { selectChoice } from '../actions';
 
 import { choicesImages } from '../assets/images';
 
-import { Desktop, Mobile } from '../MediaQuery'
+import { Desktop, Mobile } from '../styles/MediaQuery'
 
+import { Question, MobileQuestion } from '../styles/Question';
+import { ChoiceList, ChoiceMobileList } from '../styles/ChoiceList';
 import { ChoiceButton, MobileChoiceButton } from '../styles/ChoiceButton';
 
 import { LazyLoadImage } from 'react-lazy-load-image-component';
-
-const Question = styled.h3({
-  margin: '5% 0',
-  textAlign: 'center',
-  fontSize: '40px',
-});
-
-const MobileQuestion = styled.h3({
-  margin: '0',
-  height: '12vh',
-  textAlign: 'center',
-  fontSize: '30px',
-});
-
-const List = styled.ul({
-  display: 'grid',
-  gridTemplateColumns: 'repeat(2, 1fr)',
-  listStyle: 'none',
-  margin: '0',
-  padding: '0',
-});
-
-const MobileList = styled.ul({
-  display: 'grid',
-  gridTemplateRows: 'repeat(2, 1fr)',
-  height: '60vh',
-  listStyle: 'none',
-  margin: '0',
-  padding: '0',
-});
 
 const Item = styled.li({
   margin: '5%',
@@ -62,7 +34,7 @@ export default function ChoiceContainer({
     <div>
       <Desktop>
         <Question>{question}</Question>
-        <List>
+        <ChoiceList>
           {choices.map((choice) => (
             <Item key={choice.id}>
               <ChoiceButton
@@ -79,11 +51,11 @@ export default function ChoiceContainer({
               </ChoiceButton>
             </Item>
           ))}
-        </List>
+        </ChoiceList>
       </Desktop>
       <Mobile>
         <MobileQuestion>{question}</MobileQuestion>
-        <MobileList>
+        <ChoiceMobileList>
           {choices.map((choice) => (
             <Item key={choice.id}>
               <MobileChoiceButton
@@ -100,7 +72,7 @@ export default function ChoiceContainer({
               </MobileChoiceButton>
             </Item>
           ))}
-        </MobileList>
+        </ChoiceMobileList>
       </Mobile>
     </div>
   );
