+++ src/pages/ActivityPage.jsx
@@ -2,119 +2,17 @@ import React from 'react';
 
 import { useHistory, useParams } from 'react-router-dom';
 
-import { Desktop, Mobile } from '../MediaQuery';
-
-import styled from '@emotion/styled';
+import { Desktop, Mobile } from '../styles/MediaQuery';
 
 import resultData from '../data/resultData';
 
 import { vitaminsImages } from '../assets/images';
 
 import Back from '../styles/Back';
-
-const Container = styled.div({
-  display: 'grid',
-  gridTemplateColumns: 'repeat(2, 1fr)',
-  height: '70vh',
-});
-
-const MobileContainer = styled.div({
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'start',
-  textAlign: 'center',
-  '& h2': {
-    color: '#FFB61D',
-    fontSize: '30px',
-    margin: '0',
-  },
-  '& img': {
-    width: '250px',
-    height: '200px',
-    objectFit: 'cover',
-    margin: '5% auto',
-    borderRadius: '10px',
-    boxShadow: '0 0 10px 5px rgba(0, 0, 0, 0.2)',
-  },
-  '& p': {
-    margin: '5%',
-    fontSize: '20px',
-  },
-});
-
-const WrapperLeft = styled.div({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  margin: '5vh 0',
-  '& h2': {
-    color: '#FFB61D',
-    fontSize: '40px',
-    marginBottom: '20px',
-  },
-  '& div': {
-    height: '90%',
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    paddingBottom: '5vh',
-  },
-  '& img': {
-    width: '35vw',
-    maxWidth: '400px',
-    height: '35vw',
-    maxHeight: '400px',
-    objectFit: 'cover',
-    borderRadius: '10px',
-  },
-});
-
-const WrapperRight = styled.div({
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'center',
-  margin: '5vh 0',
-  padding: '15px',
-  '& > p': {
-    paddingLeft: '10px',
-    color: '#323232',
-    fontSize: '22px',
-    lineHeight: '35px',
-  },
-});
-
-const List = styled.ul({
-  listStyle: 'none',
-  textAlign: 'center',
-  backgroundColor: 'rgba(242, 201, 76, 0.2)',
-  borderRadius: '10px',
-  padding: '15px',
-  fontSize: '25px',
-});
-
-const MobileList = styled.ul({
-  listStyle: 'none',
-  textAlign: 'center',
-  backgroundColor: 'rgba(242, 201, 76, 0.2)',
-  borderRadius: '10px',
-  margin: '5%',
-  padding: '10px',
-  fontSize: '20px',
-});
-
-const Item = styled.li({
-  backgroundColor: 'rgba(255, 255, 255, 0.8)',
-  margin: '10px',
-  padding: '10px',
-  borderRadius: '10px',
-});
-
-const MobileItem = styled.li({
-  backgroundColor: 'rgba(255, 255, 255, 0.8)',
-  margin: '10px',
-  padding: '10px',
-  borderRadius: '10px',
-});
+import { ActivityContainer, ActivityMobileContainer } from '../styles/ActivityContainer';
+import { WrapperLeft, WrapperRight } from '../styles/Wrapper';
+import { RelatedList, RelatedMobileList } from '../styles/RelatedList.jsx';
+import { RelatedItem, RelatedMobileItem } from '../styles/RelatedItem';
 
 const ACTIVITY_COUNT = 6;
 
@@ -136,7 +34,7 @@ export default function ActivityPage() {
     <div>
       <Desktop>
         <Back><button type="button" data-testid="back" onClick={handleClick}><i className="fas fa-chevron-left" /></button></Back>
-        <Container>
+        <ActivityContainer>
           <WrapperLeft>
             <h2>{text}</h2>
             <div><img src={vitaminsImages[activityId.id]} alt="" /></div>
@@ -145,38 +43,38 @@ export default function ActivityPage() {
             <p>
               {detail}
             </p>
-            <List>
+            <RelatedList>
               연관 비타민
             {intro.map(({
               id, text,
             }) => (
-              <Item key={id}>
+              <RelatedItem key={id}>
                 {text}
-              </Item>
+              </RelatedItem>
             ))}
-            </List>
+            </RelatedList>
           </WrapperRight>
-        </Container>
+        </ActivityContainer>
       </Desktop>
       <Mobile>
         <Back><button type="button" data-testid="back" onClick={handleClick}><i className="fas fa-chevron-left" /></button></Back>
-        <MobileContainer>
+        <ActivityMobileContainer>
           <h2>{text}</h2>
           <img src={vitaminsImages[activityId.id]} alt="" />
           <p>
             {detail}
           </p>
-          <MobileList>
+          <RelatedMobileList>
             연관 비타민
             {intro.map(({
             id, text,
           }) => (
-            <MobileItem key={id}>
+            <RelatedMobileItem key={id}>
               {text}
-            </MobileItem>
+            </RelatedMobileItem>
           ))}
-          </MobileList>
-        </MobileContainer>
+          </RelatedMobileList>
+        </ActivityMobileContainer>
       </Mobile>
     </div>
   );
