+++ src/styles/More.jsx
@@ -1,0 +1,20 @@
+import styled from '@emotion/styled';
+
+const More = styled.div({
+  margin: '0',
+  textAlign: 'center',
+  padding: '30px 0',
+  '& a': {
+    color: 'white',
+    textDecoration: 'none',
+    fontSize: '30px',
+    backgroundColor: '#F2C94C',
+    padding: '10px 18px',
+    borderRadius: '10px',
+    '&:hover': {
+      backgroundColor: '#FFB61D',
+    },
+  },
+});
+
+export default More;
