+++ src/styles/Question.jsx
@@ -1,0 +1,16 @@
+import styled from '@emotion/styled';
+
+const Question = styled.h3({
+  margin: '5% 0',
+  textAlign: 'center',
+  fontSize: '40px',
+});
+
+const MobileQuestion = styled.h3({
+  margin: '0',
+  height: '12vh',
+  textAlign: 'center',
+  fontSize: '30px',
+});
+
+export { Question, MobileQuestion };
