+++ src/styles/RelatedItem.jsx
@@ -1,0 +1,17 @@
+import styled from '@emotion/styled';
+
+const RelatedItem = styled.li({
+  backgroundColor: 'rgba(255, 255, 255, 0.8)',
+  margin: '10px',
+  padding: '10px',
+  borderRadius: '10px',
+});
+
+const RelatedMobileItem = styled.li({
+  backgroundColor: 'rgba(255, 255, 255, 0.8)',
+  margin: '10px',
+  padding: '10px',
+  borderRadius: '10px',
+});
+
+export { RelatedItem, RelatedMobileItem };
