+++ src/styles/RelatedList.jsx
@@ -1,0 +1,22 @@
+import styled from '@emotion/styled';
+
+const RelatedList = styled.ul({
+  listStyle: 'none',
+  textAlign: 'center',
+  backgroundColor: 'rgba(242, 201, 76, 0.2)',
+  borderRadius: '10px',
+  padding: '15px',
+  fontSize: '25px',
+});
+
+const RelatedMobileList = styled.ul({
+  listStyle: 'none',
+  textAlign: 'center',
+  backgroundColor: 'rgba(242, 201, 76, 0.2)',
+  borderRadius: '10px',
+  margin: '5%',
+  padding: '10px',
+  fontSize: '20px',
+});
+
+export { RelatedList, RelatedMobileList };
