+++ src/styles/Title.jsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 
 const Title = styled.h2({
-  margin: '5% 0',
+  margin: '1% 0',
   textAlign: 'center',
   fontSize: '40px',
 });
