+++ src/styles/globalStyle.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import { Global, css } from '@emotion/react';
 
 export const GlobalStyle = () => (
